requirements.txt
pandas>=2.0.0
python-dateutil>=2.8.2
matplotlib>=3.7.0
pyyaml>=6.0
